Synchronize LTS after Initcond setup.
