+++ src/Initializer/InitialFieldProjection.cpp
@@ -40,6 +40,8 @@
 
 #include "InitialFieldProjection.h"
 
+#include "Initializer/tree/LTSSync.hpp"
+
 #include <Numerical_aux/Quadrature.h>
 #include <Numerical_aux/BasisFunction.h>
 #include <Numerical_aux/Transformation.h>
@@ -120,4 +122,9 @@ void seissol::initializers::projectInitialField(std::vector<std::unique_ptr<phys
 #ifdef _OPENMP
   }
 #endif
+
+  seissol::initializer::synchronize(lts.dofs);
+  if (kernels::size<tensor::Qane>() > 0) {
+    seissol::initializer::synchronize(lts.dofsAne);
+  }
 }
